requirements.txt
pillow
configparser